MAJOR REFACTOR
This is a major overhaul of my dotfile setup.  In it are 3 major changes to mysetup:

1.) Rakefile

Now the install task will verify that oh-my-zsh, pathogen, and vundle
are installed on the box and if any of them are missing the task will
install them.

 2.) Aliases & Functions

Next, I've ditched the zsh alias and function files in favor of
oh-my-zsh plugins.

3.) Themes and Colors

Lastly, I've ditched the dark theme for a lighter github color scheme. I
also changed my tmux terminal to lighter color scheme too.