RAKE FILE REFACTOR
This change mostly consists of an overhaul of the Rakefile.  I was not
fond of the structure.

I also made a slight change to my custom git log (g la).  Changed the
color from yellow to red so that it is easier to read in my new light
colorscheme terminal.